sesai-backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
pymysql==1.1.0
cryptography==42.0.0
alembic==1.13.1

# OpenAI
openai==1.12.0

# Google APIs
google-api-python-client==2.116.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-generativeai>=0.3.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Configuration
pydantic-settings==2.1.0
python-dotenv==1.0.0

# File Processing
PyPDF2==3.0.1
Pillow==10.2.0
python-magic==0.4.27

# Utilities
httpx==0.26.0
aiofiles==23.2.1

# Testing (optional)
pytest==7.4.4
pytest-asyncio==0.23.4

